+++ crankshaft-engine/src/lib.rs
@@ -1,13 +1,7 @@
 //! The engine that powers Crankshaft.
 
-use std::time::Duration;
-
 use crankshaft_config::backend::Config;
-use futures::StreamExt;
-use futures::stream::FuturesUnordered;
 use indexmap::IndexMap;
-use indicatif::ProgressBar;
-use indicatif::ProgressStyle;
 use tracing::debug;
 
 pub mod service;
@@ -26,18 +20,15 @@ use crate::service::runner::TaskHandle;
 /// be displayed directly to the user.
 ///
 /// In cases where an error may be recoverable throughout this crate, a
-/// different error type may be returned (as it will always be coercable to it's
+/// different error type may be returned (as it will always be coercible to it's
 /// [`anyhow`] equivalent for display).
 pub type Result<T> = eyre::Result<T>;
-
-/// Runners stored within the engine.
-type Runners = IndexMap<String, Runner>;
 
 /// A workflow execution engine.
 #[derive(Debug, Default)]
 pub struct Engine {
     /// The task runner(s).
-    runners: Runners,
+    runners: IndexMap<String, Runner>,
 }
 
 impl Engine {
@@ -54,11 +45,11 @@ impl Engine {
         self.runners.keys().map(|key| key.as_ref())
     }
 
-    /// Submits a [`Task`] to be executed.
+    /// Spawns a [`Task`] to be executed.
     ///
     /// A [`Handle`] is returned, which contains a channel that can be awaited
     /// for the result of the job.
-    pub fn submit(&self, name: impl AsRef<str>, task: Task) -> TaskHandle {
+    pub fn spawn(&self, name: impl AsRef<str>, task: Task) -> Result<TaskHandle> {
         let name = name.as_ref();
         let backend = self
             .runners
@@ -73,7 +64,7 @@ impl Engine {
             name
         );
 
-        backend.submit(task)
+        backend.spawn(task)
     }
 
     /// Starts an instrumentation loop.
@@ -91,34 +82,5 @@ impl Engine {
                 tokio::time::sleep(Duration::from_millis(delay_ms)).await;
             }
         });
-    }
-
-    /// Runs all of the tasks scheduled in the engine.
-    pub async fn run(self) {
-        let mut futures = FuturesUnordered::new();
-
-        for (_, runner) in self.runners {
-            futures.extend(runner.tasks());
-        }
-
-        let task_completion_bar = ProgressBar::new(futures.len() as u64);
-        task_completion_bar.set_style(
-            ProgressStyle::with_template(
-                "{spinner:.cyan/blue} [{elapsed_precise}] [{wide_bar:.cyan/blue}] \
-                 {pos:>7}/{len:7} {msg}",
-            )
-            .unwrap()
-            .progress_chars("#>-"),
-        );
-
-        let mut count = 1;
-        task_completion_bar.inc(0);
-        task_completion_bar.enable_steady_tick(Duration::from_millis(100));
-
-        while (futures.next().await).is_some() {
-            task_completion_bar.set_message(format!("task #{}", count));
-            task_completion_bar.inc(1);
-            count += 1;
-        }
     }
 }