+++ examples/src/lsf/main.rs
@@ -4,6 +4,9 @@
 //!
 //! `cargo run --release --example lsf`
 
+use std::env::current_dir;
+use std::time::Duration;
+
 use clap::Parser;
 use crankshaft::Config;
 use crankshaft::Engine;
@@ -12,8 +15,11 @@ use crankshaft::engine::task::Execution;
 use eyre::Context as _;
 use eyre::ContextCompat as _;
 use eyre::Result;
+use futures::StreamExt;
+use futures::stream::FuturesUnordered;
+use indicatif::ProgressBar;
+use indicatif::ProgressStyle;
 use nonempty::NonEmpty;
-use tracing::info;
 use tracing_subscriber::EnvFilter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::layer::SubscriberExt as _;
@@ -52,7 +58,7 @@ const CONFIG: &str = r#"backends:
               -o ~{cwd}/stdout.lsf
               -e ~{cwd}/stderr.lsf
               -R "rusage[mem=~{ram_mb}] span[hosts=~{hosts}]"
-              ~{shell}
+              ~{command}
 "#;
 
 /// Starting point for task execution.
@@ -65,28 +71,65 @@ async fn run(args: Args) -> Result<()> {
 
     let engine = Engine::default().with(config).await?;
 
-    let executions = NonEmpty::new(
-        Execution::builder()
-            .workdir(".")
-            .image("ubuntu")
-            .args((String::from("echo"), vec![String::from("'hello, world!'")]))
-            .build(),
-    );
-
     let task = Task::builder()
-        .name("my-example-task")
         .description("a longer description")
-        .executions(executions)
+        .executions(NonEmpty::new(
+            Execution::builder()
+                .work_dir(
+                    current_dir()
+                        .expect("a current working directory")
+                        .display()
+                        .to_string(),
+                )
+                .image("ubuntu")
+                .program("echo")
+                .args([String::from("hello, world!")])
+                .build(),
+        ))
         .build();
 
-    let receivers = (0..args.n_jobs)
-        .map(|_| engine.submit("lsf", task.clone()).callback)
-        .collect::<Vec<_>>();
+    let mut tasks = (0..args.n_jobs)
+        .map(|_| Ok(engine.spawn("lsf", task.clone())?.wait()))
+        .collect::<Result<FuturesUnordered<_>>>()?;
 
-    engine.run().await;
+    let progress = ProgressBar::new(tasks.len() as u64);
+    progress.set_style(
+        ProgressStyle::with_template(
+            "{spinner:.cyan/blue} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos:>7}/{len:7} \
+             {msg}",
+        )
+        .unwrap()
+        .progress_chars("#>-"),
+    );
 
-    for rx in receivers {
-        info!(reply = ?rx.await.unwrap());
+    progress.enable_steady_tick(Duration::from_millis(100));
+
+    let mut results = Vec::new();
+    while let Some(result) = tasks.next().await {
+        let failed = result.is_err();
+        results.push(result);
+
+        progress.set_message(format!(
+            "task #{num} {status}",
+            num = results.len(),
+            status = if failed { "failed" } else { "completed" }
+        ));
+        progress.inc(1);
+    }
+
+    drop(progress);
+
+    for (i, result) in results.into_iter().enumerate() {
+        match result {
+            Ok(output) => println!(
+                "task #{num} {status}, stdout: {stdout:?}, stderr: {stderr:?}",
+                num = i + 1,
+                status = output.first().status,
+                stdout = std::str::from_utf8(&output.first().stdout).unwrap_or("<not UTF-8>"),
+                stderr = std::str::from_utf8(&output.first().stderr).unwrap_or("<not UTF-8>")
+            ),
+            Err(e) => println!("task #{num} failed: {e:#}", num = i + 1),
+        }
     }
 
     Ok(())