+++ examples/src/tes/main.rs
@@ -9,6 +9,9 @@
 //! cargo run --release --example tes <URL>
 //! ```
 
+use std::env::current_dir;
+use std::time::Duration;
+
 use base64::Engine as _;
 use base64::engine::general_purpose::STANDARD;
 use clap::Parser;
@@ -19,8 +22,12 @@ use crankshaft::config::backend::tes::http;
 use crankshaft::engine::Task;
 use crankshaft::engine::task::Execution;
 use eyre::Result;
+use eyre::bail;
+use futures::StreamExt;
+use futures::stream::FuturesUnordered;
+use indicatif::ProgressBar;
+use indicatif::ProgressStyle;
 use nonempty::NonEmpty;
-use tracing::info;
 use tracing_subscriber::EnvFilter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::layer::SubscriberExt as _;
@@ -56,7 +63,7 @@ async fn run(args: Args) -> Result<()> {
     let password = std::env::var(PASSWORD_ENV).ok();
 
     if (username.is_some() && password.is_none()) || (username.is_none() && password.is_some()) {
-        panic!("both username and password must be provided for authentication");
+        bail!("both username and password must be provided for authentication");
     }
 
     let mut http_config = http::Config::default();
@@ -76,31 +83,68 @@ async fn run(args: Args) -> Result<()> {
 
     let engine = Engine::default().with(config).await?;
 
-    let executions = NonEmpty::new(
-        Execution::builder()
-            .workdir(".")
-            .image("ubuntu")
-            .args((String::from("echo"), vec![String::from("'hello, world!'")]))
-            .build(),
-    );
-
     let task = Task::builder()
-        .name("my-example-task")
         .description("a longer description")
-        .executions(executions)
+        .executions(NonEmpty::new(
+            Execution::builder()
+                .work_dir(
+                    current_dir()
+                        .expect("a current working directory")
+                        .display()
+                        .to_string(),
+                )
+                .image("ubuntu")
+                .program("echo")
+                .args([String::from("hello, world!")])
+                .build(),
+        ))
         .build();
-
-    let receivers = (0..args.n_jobs)
-        .map(|_| engine.submit("tes", task.clone()).callback)
-        .collect::<Vec<_>>();
 
     #[cfg(tokio_unstable)]
     Engine::start_instrument(3000);
 
-    engine.run().await;
+    let mut tasks = (0..args.n_jobs)
+        .map(|_| Ok(engine.spawn("tes", task.clone())?.wait()))
+        .collect::<Result<FuturesUnordered<_>>>()?;
 
-    for rx in receivers {
-        info!(reply = ?rx.await.unwrap());
+    let progress = ProgressBar::new(tasks.len() as u64);
+    progress.set_style(
+        ProgressStyle::with_template(
+            "{spinner:.cyan/blue} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos:>7}/{len:7} \
+             {msg}",
+        )
+        .unwrap()
+        .progress_chars("#>-"),
+    );
+
+    progress.enable_steady_tick(Duration::from_millis(100));
+
+    let mut results = Vec::new();
+    while let Some(result) = tasks.next().await {
+        let failed = result.is_err();
+        results.push(result);
+
+        progress.set_message(format!(
+            "task #{num} {status}",
+            num = results.len(),
+            status = if failed { "failed" } else { "completed" }
+        ));
+        progress.inc(1);
+    }
+
+    drop(progress);
+
+    for (i, result) in results.into_iter().enumerate() {
+        match result {
+            Ok(output) => println!(
+                "task #{num} {status}, stdout: {stdout:?}, stderr: {stderr:?}",
+                num = i + 1,
+                status = output.first().status,
+                stdout = std::str::from_utf8(&output.first().stdout).unwrap_or("<not UTF-8>"),
+                stderr = std::str::from_utf8(&output.first().stderr).unwrap_or("<not UTF-8>")
+            ),
+            Err(e) => println!("task #{num} failed: {e:#}", num = i + 1),
+        }
     }
 
     Ok(())